+++ packages/api/utils/audio-over-socket.js
@@ -164,8 +164,10 @@ export function registerStreamForRecorder(recorder, rpcName) {
           /**
            * Notify that we are ready to process audio.
            * @event broadcaster#websocketserverreadyforaudio
+           * @deprecated will be removed in a future version
            */
           broadcaster.emit('websocketserverreadyforaudio');
+          recorder.dispatchEvent(new Event('websocketserverreadyforaudio'));
           resolve(registration);
         })
         .catch(reject);
@@ -238,7 +240,10 @@ export function prepareServerForAudio(id, recorder, rpc) {
   }).then(() => {
     // We've prepped the websocket server, now it can receive audio. Broadcast
     // that it is allowed to record.
+    // This call is deprecated and will be removed in a future version, the
+    // event on the recorder will stay.
     broadcaster.emit('websocketserverreadyforaudio');
+    recorder.dispatchEvent(new Event('websocketserverreadyforaudio'));
     return id;
   });
 }